Move ckpt convertion to tools.
