+++ mmseg/models/utils/__init__.py
@@ -1,4 +1,4 @@
-from .ckpt_convert import mit_convert, swin_convert, vit_convert
+from .ckpt_convert import swin_convert, vit_convert
 from .embed import PatchEmbed
 from .inverted_residual import InvertedResidual, InvertedResidualV3
 from .make_divisible import make_divisible
@@ -11,5 +11,5 @@ from .up_conv_block import UpConvBlock
 __all__ = [
     'ResLayer', 'SelfAttentionBlock', 'make_divisible', 'InvertedResidual',
     'UpConvBlock', 'InvertedResidualV3', 'SELayer', 'vit_convert',
-    'mit_convert', 'swin_convert', 'PatchEmbed', 'nchw_to_nlc', 'nlc_to_nchw'
+    'swin_convert', 'PatchEmbed', 'nchw_to_nlc', 'nlc_to_nchw'
 ]
